[ACR/5] Initial commit of the az_iot SDK (sas token module and tests)
