+++ sdk/iot/core/src/az_iot_sas_token.c
@@ -1,11 +1,12 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
 
-#include <stdlib.h>
-#include <stdint.h>
-#include <az_iot_sas_token.h>
 #include <az_contract_internal.h>
+#include <az_iot_sas_token.h>
 #include <az_span.h>
+
+#include <stdint.h>
+
 #include <_az_cfg.h>
 
 #define LF                      '\n'
